python-cv/requirements.txt
# ── Core API ────────────────────────────────────────────────────────────────
fastapi
uvicorn[standard]
python-multipart
python-dotenv

# ── HTTP clients ──────────────────────────────────────────────────────────────
requests
httpx

# ── Computer Vision ───────────────────────────────────────────────────────────
# opencv-python-headless is mandatory for headless servers (Render, Docker).
# Do NOT use opencv-python — it requires GUI libraries that crash on Render.
opencv-python-headless
numpy

# ── Image handling ────────────────────────────────────────────────────────────
Pillow

# ── Face Recognition and DB ──────────────────────────────────────────────────
face-recognition
dlib
pymongo

